requirements.txt
boto3==1.26.145
botocore==1.29.145
iniconfig==2.0.0
jmespath==1.0.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pluggy==1.0.0
pyspark==3.4.0
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
tzdata==2023.3
urllib3==1.26.16
